transfers:searchDone

+++ models/footballer.cpp
@@ -2,7 +2,7 @@
 #include "team.h"
 
 /*-> constructor <-*/
-Footballer::Footballer(string name, int price, string position, string teamName): footballerId(idCounter++){
+Footballer::Footballer(string name, float price, string position, string teamName): footballerId(idCounter++){
     this->Name=name;
     this->Price=price;
     this->Position=position;
@@ -19,7 +19,7 @@ string Footballer::getFootballerName(){
 string Footballer::getFootballerPosition(){
     return Position;
 };
-int Footballer::getFootballerPrice(){
+float Footballer::getFootballerPrice(){
     return Price;
 };
 int Footballer:: getGameWeekPoints(){
@@ -34,7 +34,7 @@ string Footballer:: getFootballerTeam(){
 void Footballer::setFootballerposition(string position){
     this->Position=position;
 };
-void Footballer::setFootballerPrice(int price){
+void Footballer::setFootballerPrice(float price){
     this->Price=price;
 };
 void Footballer::setFootballerTeam(int teamId){
