+++ models/footballer.h
@@ -10,7 +10,7 @@ private:
     static int idCounter;
     const int footballerId;
     string Name;
-    int Price;
+    float Price;
     string Position;
     int teamId;
     string teamName;
@@ -19,12 +19,12 @@ private:
 public:
     /*-> constructor <-*/
     Footballer();
-    Footballer(string Name,int price,string position, string teamName);
+    Footballer(string Name,float price,string position, string teamName);
 
     /*-> getters <-*/
     string getFootballerName();
     string getFootballerPosition();
-    int getFootballerPrice();
+    float getFootballerPrice();
     // Team getFootballerTeam();
     int getFootballerId();
     int getGameWeekPoints();
@@ -32,7 +32,7 @@ public:
 
     /*-> setters <-*/
     void setFootballerposition(string position);
-    void setFootballerPrice(int price);
+    void setFootballerPrice(float price);
     void setFootballerTeam(int teamId);
     void setGameWeekPoints(int points);
 };
